Move collection dropdown test coverage to isolated test scene
